requirements.txt
ffmpeg-python==0.1.18
ffpyplayer==4.2.0
future==0.17.1
mutagen==1.42.0
numpy==1.16.4
opencv-python==4.1.0.25
pyee==6.0.0
